# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MIS Backend HTTP API
  version: "1.0"
  description: |
    Specification of **actual** Micronaut application paths (prefix `/api/...`, not `/api/v1`).
    Complements the generated Swagger (`/swagger-ui`, build artifacts under `META-INF/swagger`).

    ## Authentication
    - **JWT Bearer**: `Authorization: Bearer <token>` (login response includes a `token` field).
    - **Cookie**: HttpOnly `MIS_AUTH=<token>` (set on successful `/api/auth/login` and `/api/auth/first-password-change`).
    - Anonymous: only `/api/auth/login`, `/api/auth/first-password-change`, and `GET`/`POST` under `/api/public/booking/**` (see `application.properties`).

    ## REST access control (`ApiAccessControlFilter`, `/api/**`)
    After authentication, most routes resolve to **exactly one** permission code (`x-required-permission`).
    If the code is absent (`null`), being logged in is enough. If a code is present, it must be in the employee’s effective permission set (role ± overrides via `/api/access/...`).

    Prefix → code mapping (`GET` vs everything else):
    | Prefix / condition | GET | POST, PUT, PATCH, DELETE |
    |---|---|---|
    | `/api/shell/bootstrap?kind=…` | see `kind` below | — |
    | `/api/access` | `permission.read` | `permission.write` |
    | `/api/organizations` | `organization.read` | `organization.write` |
    | `/api/branches` | `branch.read` | `branch.write` |
    | `/api/employees` | `employee.read` | `employee.write` |
...

...

...

    ## Errors
...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      x-required-permission: organization.read
      parameters:
        - $ref: "#/components/parameters/pageablePage"
        - $ref: "#/components/parameters/pageableSize"
      responses:
        "200": { description: Page }
    post:
      tags: [Organizations]
      summary: Create organization
      x-required-permission: organization.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
  /api/organizations/{id}:
    get:
      tags: [Organizations]
      summary: Organization by id
      x-required-permission: organization.read
      parameters:
        - $ref: "#/components/parameters/idUuid"
      responses:
        "200": { description: OK }
        "404": { description: Not found }
    patch:
      tags: [Organizations]
      summary: Update organization
      x-required-permission: organization.write
      parameters:
        - $ref: "#/components/parameters/idUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
    delete:
      tags: [Organizations]
      summary: Delete organization
      x-required-permission: organization.write
      parameters:
        - $ref: "#/components/parameters/idUuid"
      responses:
        "204": { description: Deleted }
...
  /api/employees/for-online-booking:
    get:
      tags: [Employees]
      summary: Employees for online booking widget (no pagination)
      x-required-permission: employee.read
      parameters:
        - name: branchId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /api/employees:
    get:
      tags: [Employees]
      summary: List employees
      x-required-permission: employee.read
      parameters:
        - $ref: "#/components/parameters/pageablePage"
        - $ref: "#/components/parameters/pageableSize"
      responses:
        "200": { description: Page }
    post:
      tags: [Employees]
      summary: Create employee
      x-required-permission: employee.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
  /api/employees/{id}:
    get:
      tags: [Employees]
      summary: Employee by id
      x-required-permission: employee.read
      parameters:
        - $ref: "#/components/parameters/idUuid"
      responses:
        "200": { description: OK }
    patch:
      tags: [Employees]
      summary: Update employee
      x-required-permission: employee.write
      parameters:
        - $ref: "#/components/parameters/idUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
    delete:
      tags: [Employees]
      summary: Delete employee
      x-required-permission: employee.write
      parameters:
        - $ref: "#/components/parameters/idUuid"
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from `token` field in `/api/auth/login` response
    cookieAuth:
      type: apiKey
      in: cookie
      name: MIS_AUTH
      description: HttpOnly cookie (alternative to Bearer)
  parameters:
    idUuid:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    pageablePage:
      name: page
      in: query
      schema: { type: integer, minimum: 0 }
      description: Page number (Micronaut Pageable)
    pageableSize:
      name: size
      in: query
      schema: { type: integer, minimum: 1 }
      description: Page size
...